refactor(sql): extract error classification into sql_errors module
Move all error-classification constants (_PERMISSION_DENIED_FRAGMENTS,
_AUTH_FAILED_FRAGMENTS, _PERMISSION_DENIED_ERROR_NUMBERS,
_AUTH_FAILED_ERROR_NUMBERS, _NOT_FOUND_ERROR_NUMBERS, _NOT_FOUND_FRAGMENTS,
_SNAPSHOT_NOT_READY_FRAGMENTS, _TRANSIENT_FRAGMENTS, _NATIVE_ERROR_RE,
_DRIVER_NOISE_RE) and functions (map_driver_error, _clean_driver_error_message,
_wrap_unmapped_driver_error, is_transient_connection_error,
is_snapshot_not_ready_error, is_auth_failed_message, _is_connect_retryable)
out of sql.py into a new leaf module sql_errors.py whose only production
dependency is fabric_dw.exceptions (plus stdlib re).

sql.py re-imports all moved names so that existing callers using
`from fabric_dw.sql import map_driver_error` (and the is_* helpers) and
any `patch("fabric_dw.sql.<name>")` test patches continue to resolve
without modification. The __all__ list in sql.py is unchanged.

Part of #828
